Add live preview of rendered slides for authoring
The slide content can be viewed by using the Spectacle-CLI with the
slide content in Markdown and the slide theme in a .js file.

To start up the development slide renderer, you will need to first
install the 'spectacle-cli' by following the 'Getting Started' section
of the README.md. Then the author will be able to start the development
server and pass the slides and theme desired for previewing to the
'spectacle-cli' tool as flag arguments. Any changes made to the Markdown
files will be automatically converted into React JSX and injected into
the preview template after a couple of seconds.

Please refer to the README.md file for additional details.